+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Styled the ButtonPanel component
 - Styled the Button component
 - Passed "color" and "wide" props to the Button
+- Logic Folder with `calculate` and `operate`
+- Used `big.js`
+- Import `Calculate` into `App`
 
 ### Fixed
 - `calcRow` height to be 100px